Update routes and controller actions for Statements
My old routes didn't work with Rails 4, and they actually didn't make a
lot of sense. Not sure what I was thinking. Really what I have is a new
statement form and a create statement action which just directly renders
it's result since it's not being saved in the db and given an id to be
pulled up later.

I haven't tested this change, but I'm guessing it will work. Need to
test.