Use secrets.yml
Since currently the only secret is the SECRET_KEY_BASE and since Rails
4.1 recommends putting that config in secrets.yml, I went ahead and
removed figaro and application.example.yml. Figaro could still be useful
if I wanted to store the production SECRET_KEY_BASE in a gitignored
application.yml and use figaro to set the environment variable in
production, but I think I can handle that manually for now.