Only show flash messages for warning/error cases
I Removed all the notice type flash messages that just tell you that
whatever you just did was successful. I left in the ones that warned the
user about a problem or that they were not allowed to do something, etc.

I also fixed the code that determines the class names for the alerts,
and updated the alert html to conform to bootstrap 3 norms. Also added
some css to put some margin between the navbar and the alert, or between
alert and alert if there was some way for there to be more than one.