User can add/edit/update notes and title for batches
When a new batch is created the user can give it an optional title and
enter in any notes they would like to have associated with the batch.
Then, at any time while viewing the batch contributions for that batch
a user may click 'Edit Title & Notes' and modify those values as
desired.

The notes are displayed at the bottom of the batch contributions index
view as well as on the bottom of the batch report PDF.

Did a little bit of refactoring of the batches controller.