Init date fields again after javascript updates the DOM
I moved all the code needed to set up the date fields into a function
and exposed the function on the global app variable, Giving. Then I call
that function whenever necessary to initialize newly added fields.

I also added code to set the focus to the beginning of the form (the
contributor name field), as the autofocus attribute doesn't seem to take
effect when just those fields are being replaced in the DOM with js.