Refactor contributor select html and javascript
The html and javascript for the contributor select portion of the forms
were exactly the same between the batch contributions and contributions
forms and inline forms. So, I extracted the html and the js.erb into a
couple of shared partials:

shared/_contributor_select.html.haml - the html for the control
shared/_contributors_javascript.html.erb - the js that assigns the
Giving.contributors variable to an array of json objects that the
typeahead uses.

The idea with separating the javascript into its own partial is that I
intend to implement caching on that so that it doesn't have to hit the
database every time unless a new contributor has been added or an
existing one has been modified. I haven't added the caching yet, but it
will be trivial.

I then moved the javascript code that does not need to be evaluated with
ruby first into app/assets/javascripts/contributors_select.js.coffee.
That is, the code to initialize the typeahead to select the contributor.
I put that code in a function that I tacked on to the Giving global so
that the js partials can call it to re-initialize the control as
necessary.

This is all an attempt to DRY things up in preparation for adding some
additional functionality to the contributor select control, specifically
the ability to create a contributor inline while adding a contribution
for them.