Submit contributor name instead of id in contribution form
Switched the lookup of the id to the server side. This fixes the issue
where a user selects a contributor and then goes back and changes the
name in the name field, which wasn't resetting the hidden id field.

Now whatever is entered in the name field gets submitted and the id is
looked up from the name. (Names are now unique). So, currently an
invalid name will result in no contributor selected.

This is mainly in preparation for a new option which will allow a new
contributor to be created if the user so desires.