Allow creation of contributors from the contribution forms
If the contributor name entered does not match an existing contributor
name, an alert is displayed asking the user if they want to create the
new contributor and letting them know that they will need to go in later
and fill out all the rest of the details. If the user clicks ok it will
create the user via ajax and continue on to the date field. If the user
clicks cancel it clears the contributor name field and sets the focus
back on that field.

I ended up updating the typeahead that I was using to the bootstrap3
typeahead. There are a few small differences. Since I am no longer using
the contributor id field, Giving.contributors, which is the source of
the typeahead is just an array of contributor names rather than objects.
One thing to note is that by default for some reason the bootstrap3
typeahead has a built-in 250ms delay, which was really annoying. So I
changed it to 5ms.

I am also using an alternative to the standard js alert called Sweet
Alert. It is asynchronous so the result is handled in a callback, and
the code in the callback is actually executed before the modal is
completely closed, so I had to use setTimeout to get the code to set the
focus to the correct field to work correctly.