Make contributor name field larger, add placeholder to date field
I noticed many contributor names (particularly couples) were
considerably too long to fit in the field in the inline form as it was
previously sized. So I increased it to a 3-width column instead 2-width,
and decreased the size of the Add button to make room.

There was also some confusion on the date format exepected or allowed,
so I added a placeholder of mm/dd/yyyy. The /yyyy part is optional, but
I can let the users know about that.