Trim the name entered AND the name stored in the database on compare
There was a case where a contributor name was stored in the database
with a trailing space, and the code to check to see if the name entered
was a known contributor failed to detect it because it was trimming the
input, but not the name in the Giving.contributors list retreived from
the database. Now it trims and downcases both before comparing.