Allow all users to edit their own profile (username and password)
Added a profile route with edit and update actions. Changed the logged
in menu item to a dropdown where the root of it is the logged in user's
username, and the menu items are Edit Profile and Logout. If the user
goes to Edit Profile they can edit their username or password if
desired.