Add IndividualStatementsZipFile class
As the name hopefully indicates, this class is used to create a zip file
containing pdf statements for individual contributors. That is, one pdf
per contributor. It uses a Tempfile internally, so it should be cleaned
up eventually by the OS, but it is recommended to delete it when you're
done with it.