Add option of zip file of individual statements
Now when a user wants to generate statements they can choose to generate
one large PDF that includes all contributor statements (ideal for
printing) or to generate a zip file with individual PDFS inside for each
contributor.

Also added error handling for the case where there are no contributions
for the year that the user selected. In that case it will just render a
flash warning instead of generating an empty file.